+++ src/blob/handlers/ServiceHandler.ts
@@ -11,7 +11,7 @@ import {
   EMULATOR_ACCOUNT_ISHIERARCHICALNAMESPACEENABLED,
   EMULATOR_ACCOUNT_KIND,
   EMULATOR_ACCOUNT_SKUNAME,
-  HeaderConstants,
+  HeaderConstants
 } from "../utils/constants";
 import BaseHandler from "./BaseHandler";
 import IAccountDataStore from "../../common/IAccountDataStore";
@@ -23,7 +23,7 @@ import { Readable } from "stream";
 import { OAuthLevel } from "../../common/models";
 import { BEARER_TOKEN_PREFIX } from "../../common/utils/constants";
 import { decode } from "jsonwebtoken";
-import { getUserDelegationKeyValue } from "../utils/utils"
+import { getUserDelegationKeyValue } from "../utils/utils";
 
 /**
  * ServiceHandler handles Azure Storage Blob service related requests.
@@ -32,8 +32,10 @@ import { getUserDelegationKeyValue } from "../utils/utils"
  * @class ServiceHandler
  * @implements {IHandler}
  */
-export default class ServiceHandler extends BaseHandler
-  implements IServiceHandler {
+export default class ServiceHandler
+  extends BaseHandler
+  implements IServiceHandler
+{
   protected disableProductStyle?: boolean;
 
   constructor(
@@ -117,7 +119,6 @@ export default class ServiceHandler extends BaseHandler
     return response;
   }
 
-
   public async submitBatch(
     body: NodeJS.ReadableStream,
     contentLength: number,
@@ -125,16 +126,25 @@ export default class ServiceHandler extends BaseHandler
     options: Models.ServiceSubmitBatchOptionalParams,
     context: Context
   ): Promise<Models.ServiceSubmitBatchResponse> {
-    const blobBatchHandler = new BlobBatchHandler(this.accountDataStore, this.oauth,
-      this.metadataStore, this.extentStore, this.logger, this.loose, this.disableProductStyle);
+    const blobBatchHandler = new BlobBatchHandler(
+      this.accountDataStore,
+      this.oauth,
+      this.metadataStore,
+      this.extentStore,
+      this.logger,
+      this.loose,
+      this.disableProductStyle
+    );
 
-    const batchResponse = await blobBatchHandler.submitBatch(body,
+    const batchResponse = await blobBatchHandler.submitBatch(
+      body,
       "",
       context.request!,
-      context);
+      context
+    );
 
     const responseBody = new Readable();
-    responseBody.push(batchResponse.reponseBody);
+    responseBody.push(batchResponse.responseBody);
     responseBody.push(null);
 
     // No client request id defined in batch response, should refine swagger and regenerate from it.
@@ -172,8 +182,8 @@ export default class ServiceHandler extends BaseHandler
     const body = blobCtx.request!.getBody();
     const parsedBody = await parseXML(body || "");
     if (
-      !Object.hasOwnProperty.bind(parsedBody)('cors') &&
-      !Object.hasOwnProperty.bind(parsedBody)('Cors')
+      !Object.hasOwnProperty.bind(parsedBody)("cors") &&
+      !Object.hasOwnProperty.bind(parsedBody)("Cors")
     ) {
       storageServiceProperties.cors = undefined;
     }
@@ -243,7 +253,8 @@ export default class ServiceHandler extends BaseHandler
     }
 
     if (properties.defaultServiceVersion === undefined) {
-      properties.defaultServiceVersion = this.defaultServiceProperties.defaultServiceVersion;
+      properties.defaultServiceVersion =
+        this.defaultServiceProperties.defaultServiceVersion;
     }
 
     if (properties.staticWebsite === undefined) {
@@ -264,7 +275,6 @@ export default class ServiceHandler extends BaseHandler
     options: Models.ServiceGetStatisticsOptionalParams,
     context: Context
   ): Promise<Models.ServiceGetStatisticsResponse> {
-
     if (!context.context.isSecondary) {
       throw StorageErrorFactory.getInvalidQueryParameterValue(
         context.contextId
@@ -328,7 +338,7 @@ export default class ServiceHandler extends BaseHandler
     // TODO: Need update list out container lease properties with ContainerHandler.updateLeaseAttributes()
     const serviceEndpoint = `${request.getEndpoint()}/${accountName}`;
     const res: Models.ServiceListContainersSegmentResponse = {
-      containerItems: containers[0].map(item => {
+      containerItems: containers[0].map((item) => {
         return {
           ...item,
           metadata: includeMetadata ? item.metadata : undefined
@@ -357,7 +367,8 @@ export default class ServiceHandler extends BaseHandler
       skuName: EMULATOR_ACCOUNT_SKUNAME,
       accountKind: EMULATOR_ACCOUNT_KIND,
       date: context.startTime!,
-      isHierarchicalNamespaceEnabled: EMULATOR_ACCOUNT_ISHIERARCHICALNAMESPACEENABLED,
+      isHierarchicalNamespaceEnabled:
+        EMULATOR_ACCOUNT_ISHIERARCHICALNAMESPACEENABLED,
       version: BLOB_API_VERSION
     };
     return response;
@@ -392,7 +403,7 @@ export default class ServiceHandler extends BaseHandler
       undefined,
       options.where,
       options.maxresults,
-      marker,
+      marker
     );
 
     const serviceEndpoint = `${request.getEndpoint()}/${accountName}`;